Create page to claculate full years entitlement
Add NHS service manual CSS file to give the page NHS styling.
Create HTML page that contains radio buttons for length of service,
input field for contracted hours, and claculate button that gives the
results.
Create Javascript function to calculate the FTE, days entitlement, and
hours entitlement for annual leave. Add event listener to
calculate-button so that claculateLeave function runs when the button is
clicked on the webpage.