Add contract start and end dates for Employees
In index.html add Radio buttons and input fields for start and end dates of employee contracts.
Create main.css file to have custom css styling so the date input fields are initially hidden from view.
In main.js create a new function with if statements to show and hide date input fields depending on which radio button is checked. Also add eventlisteners to call the new funtion when the relevant radio button is checked.