Create JS pages for each individiaula html page
Edit HTML pages to connect them to the new JS pages created. In the JS
pages creat functions and add eventListeners to make sure they function
correctly. To make sure that depending on the options selected that they
navigate to the correct page.
Use sessionStorage to temproarily store values and options slected by
the user to get the end output.